src/graphrag/graphrag-requirements.txt
neo4j
neo4j-graphrag
neo4j-graphrag[openai]
neo4j-graphrag[generation]
neo4j-graphrag[llm]
neo4j-graphrag[retrievers]
neo4j-graphrag[types]
python-dotenv
langchain-community[vectorstores]
langchain-openai
requests
openai
pandas
termcolor
evaluate
bert_score
pydantic
scikit-learn
scipy